More aggresive `_escapeHtml`, and escape `restWireData`
`restWireData` was not being HTML escaped, which meant that if any HTML
was returned in a response (which is possible for non-RESTful HTTP requests),
that the HTML was embedding and messing up the HTML for the report page.
This change escapes any returned HTML.

In addition, `_escapeHtml` has also been made more strict, escaping
more than just `<` and `>` characters.
If this implementation is considered _too_ strict, we can modify it to
just replace the "major players" such as happens in [mustache.js](https://github.com/janl/mustache.js/blob/master/mustache.js#L60-L75)